requirements.txt
pytorch-lightning
optuna
torch
torchvision
hydra-core
hydra-optuna-sweeper